user auth

+++ routes/web.php
@@ -11,35 +11,35 @@ Route::get('/', [ListingController::class, 'index']);
 
 
 //Show create form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 //Store Listing Data
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 //Show Edit Form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 //Update Listing
-Route::put('listings/{listing}', [ListingController::class, 'update']);
+Route::put('listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 //Delete Listing
-Route::delete('listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 
 //Single Listing
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 //Show register form
-Route::get('/register', [UserController::class, 'signup']);
+Route::get('/register', [UserController::class, 'signup'])->middleware('guest');
 
 //Create a new user
 Route::post('/users', [UserController::class, 'store']);
 
 //Logout user
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 //show login form
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 //login
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
